Rename route file
