+++ src/routes/index.js
@@ -9,7 +9,7 @@ import ProblemSolving from "../containers/ProblemSolving/ProblemSolving";
 import Statistics from "../containers/Statistics/Statistics";
 import Javascript from "../containers/Javascript/Javascript";
 import Code from "../containers/Code/Code";
-import { ProtectedRoute,GuestRoute } from "./ProtectedRoute";
+import { ProtectedRoute, GuestRoute } from "./CustomRoute";
 import {
   HOME,
   SINGUP,
